requirements.txt
Flask~=3.0.2
openai~=1.11.0
python-dotenv~=1.0.0
langchain~=0.0.349
langchain_openai
chromadb~=0.3.29
qianfan
